+++ Google.Cloud.EntityFrameworkCore.Spanner.IntegrationTests/Model/SpannerSampleDbContext.cs
@@ -15,6 +15,7 @@
 using Google.Cloud.EntityFrameworkCore.Spanner.Metadata;
 using Microsoft.EntityFrameworkCore;
 using System.Linq;
+using Newtonsoft.Json;
 
 namespace Google.Cloud.EntityFrameworkCore.Spanner.IntegrationTests.Model
 {
@@ -209,9 +210,20 @@ namespace Google.Cloud.EntityFrameworkCore.Spanner.IntegrationTests.Model
                 entity.Property(e => e.Code).HasMaxLength(10);
 
                 entity.Property(e => e.Name).HasMaxLength(100);
+                entity.OwnsMany(e => e.Descriptions, builder =>
+                {
+                    builder.ToJson();
+                });
             });
             
-            modelBuilder.Entity<TicketSales>();
+            modelBuilder.Entity<TicketSales>(entity =>
+            {
+                entity.Property(e => e.Receipt)
+                    .HasConversion<string>(
+                        v => v == null ? null : JsonConvert.SerializeObject(v),
+                        v => v == null ? null : JsonConvert.DeserializeObject<Receipt>(v))
+                    .HasColumnType("JSON");
+            });
         }
     }
 }