+++ Google.Cloud.EntityFrameworkCore.Spanner.IntegrationTests/Model/VenueDescription.cs
@@ -1,11 +1,11 @@
-// Copyright 2025 Google LLC
-//
+// Copyright 2025 Google Inc. All Rights Reserved.
+// 
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
+// 
+//     http://www.apache.org/licenses/LICENSE-2.0
+// 
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -14,14 +14,10 @@
 
 namespace Google.Cloud.EntityFrameworkCore.Spanner.IntegrationTests.Model;
 
-public class TicketSales
+public class VenueDescription
 {
-    public long Id { get; set; }
-
-    public string CustomerName { get; set; }
-    
-    /// <summary>
-    /// Receipt is stored as a JSON column in the database.
-    /// </summary>
-    public Receipt? Receipt { get; set; }
+    public string Category { get; set; }
+    public string Description { get; set; }
+    public long Capacity { get; set; }
+    public bool Active { get; set; }
 }