+++ Google.Cloud.EntityFrameworkCore.Spanner.Tests/EntityFrameworkUsingMutationsMockServerTests.cs
@@ -174,6 +174,41 @@ namespace Google.Cloud.EntityFrameworkCore.Spanner.Tests
         }
 
         [Fact]
+        public async Task InsertVenue()
+        {
+            await using var db = new MockServerSampleDbContextUsingMutations(ConnectionString);
+            db.Venues.Add(new Venues
+            {
+                Code = "C1",
+                Name = "Concert Hall",
+                Descriptions = [
+                    new() { Category = "Concert Hall", Description = "Big concert hall", Capacity = 1000, Active = true },
+                    new() { Category = "Hall", Description = "Big hall", Capacity = 1000, Active = false },
+                ],
+            });
+            var updateCount = await db.SaveChangesAsync();
+
+            Assert.Equal(1L, updateCount);
+            Assert.Single(_fixture.SpannerMock.Requests.OfType<BeginTransactionRequest>());
+            Assert.Collection(
+                _fixture.SpannerMock.Requests.OfType<CommitRequest>(),
+                request => {
+                    Assert.Single(request.Mutations);
+                    var mutation = request.Mutations[0];
+                    Assert.Equal(Mutation.OperationOneofCase.Insert, mutation.OperationCase);
+                    Assert.Equal("Venues", mutation.Insert.Table);
+                    var row = mutation.Insert.Values[0];
+                    var cols = mutation.Insert.Columns;
+                    Assert.Equal("C1", row.Values[cols.IndexOf("Code")].StringValue);
+                    Assert.Equal("Concert Hall", row.Values[cols.IndexOf("Name")].StringValue);
+                    Assert.Equal(
+                        "[{\"Active\":true,\"Capacity\":1000,\"Category\":\"Concert Hall\",\"Description\":\"Big concert hall\"},{\"Active\":false,\"Capacity\":1000,\"Category\":\"Hall\",\"Description\":\"Big hall\"}]",
+                        row.Values[cols.IndexOf("Descriptions")].StringValue);
+                }
+            );
+        }
+
+        [Fact]
         public async Task InsertSingerInTransaction()
         {
             await using var db = new MockServerSampleDbContextUsingMutations(ConnectionString);
@@ -767,9 +802,9 @@ namespace Google.Cloud.EntityFrameworkCore.Spanner.Tests
             await using var db = new MockServerSampleDbContextUsingMutations(ConnectionString);
             var transaction = await db.Database.BeginTransactionAsync();
             db.TicketSales.AddRange([
-                new TicketSales { CustomerName = "New Customer1"},
-                new TicketSales { CustomerName = "New Customer2"},
-                new TicketSales { CustomerName = "New Customer3"},
+                new TicketSales { CustomerName = "New Customer1", Receipt = new Receipt {Date = new DateOnly(2025, 9, 1), Number = "1"}},
+                new TicketSales { CustomerName = "New Customer2", Receipt = new Receipt {Date = new DateOnly(2025, 9, 1), Number = "2"}},
+                new TicketSales { CustomerName = "New Customer3", Receipt = new Receipt {Date = new DateOnly(2025, 9, 1), Number = "3"}},
             ]);
             var updateCount = await db.SaveChangesAsync();
             await transaction.CommitAsync();
@@ -789,10 +824,12 @@ namespace Google.Cloud.EntityFrameworkCore.Spanner.Tests
         private static void ValidateInsertTicketSaleMutation(Mutation mutation, int index)
         {
             Assert.Equal(Mutation.OperationOneofCase.Insert, mutation.OperationCase);
-            Assert.Single(mutation.Insert.Columns);
+            Assert.Equal(2, mutation.Insert.Columns.Count);
             Assert.Single(mutation.Insert.Values);
             Assert.Equal("CustomerName", mutation.Insert.Columns[0]);
+            Assert.Equal("Receipt", mutation.Insert.Columns[1]);
             Assert.Equal($"New Customer{index}", mutation.Insert.Values[0].Values[0].StringValue);
+            Assert.Equal("{\"Date\":\"2025-09-01\",\"Number\":\"" + index + "\"}", mutation.Insert.Values[0].Values[1].StringValue);
         }
 
         private string AddFindSingerResult(string sql)