+++ Google.Cloud.EntityFrameworkCore.Spanner/Storage/Internal/SpannerTypeMappingSource.cs
@@ -20,6 +20,7 @@ using System.Collections.Generic;
 using System.Data;
 using System.Text.Json;
 using System.Text.RegularExpressions;
+using Microsoft.EntityFrameworkCore.Storage.Json;
 
 namespace Google.Cloud.EntityFrameworkCore.Spanner.Storage.Internal
 {
@@ -70,6 +71,8 @@ namespace Google.Cloud.EntityFrameworkCore.Spanner.Storage.Internal
         private static readonly SpannerNumericTypeMapping s_numeric = new SpannerNumericTypeMapping();
 
         private static readonly SpannerJsonTypeMapping s_json = new SpannerJsonTypeMapping();
+        private static readonly SpannerStructuralJsonTypeMapping s_structuralJson = new SpannerStructuralJsonTypeMapping("json");
+        private static readonly SpannerStringTypeMapping s_jsonAsString = new ("JSON", sqlDbType: SpannerDbType.Json, dbType: DbType.Object);
 
         private static readonly SpannerGuidTypeMapping s_guid
             = new SpannerGuidTypeMapping("STRING(36)", DbType.String);
@@ -213,6 +216,7 @@ namespace Google.Cloud.EntityFrameworkCore.Spanner.Storage.Internal
                     {typeof(decimal), s_decimal},
                     {typeof(SpannerNumeric), s_numeric},
                     {typeof(JsonDocument), s_json},
+                    {typeof(JsonElement), s_structuralJson},
                     {typeof(uint), s_uint},
                     {typeof(bool), s_bool},
                     {typeof(SpannerDate), s_date},
@@ -329,6 +333,12 @@ namespace Google.Cloud.EntityFrameworkCore.Spanner.Storage.Internal
 
             if (storeTypeName != null)
             {
+                if (storeTypeName.Equals("JSON", StringComparison.InvariantCultureIgnoreCase) &&
+                    clrType == typeof(string))
+                {
+                    return s_jsonAsString;
+                }
+                
                 if (_storeTypeMappings.TryGetValue(storeTypeName, out var mapping)
                     || _storeTypeMappings.TryGetValue(storeTypeNameBase, out mapping))
                 {